popravljen prikaz soba, zavrene opcije dodaj,obrisi,edituj

+++ Projekat/Ehotel/DDR2/ViewModel/RoomsVM.cs
@@ -1,12 +1,17 @@
 ﻿using DDR2.Helper;
 using DDR2.HotelBaza.Models;
 using DDR2.Model;
+using DDR2.View;
 using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
+using Windows.UI.Core;
+using Windows.UI.Popups;
+using Windows.UI.Xaml.Navigation;
 
 namespace DDR2.ViewModel
 {
@@ -15,25 +20,70 @@ namespace DDR2.ViewModel
         public string Pretraga { get; set; }
         public ICommand Search { get; set; }
         public ICommand Brisi { get; set; }
+        public ICommand Edituj { get; set; }
+        public ICommand AddRoom { get; set; }
         public Soba Selektovana { get; set; }
-        public List<Soba> SobeListe { get; set; }
+        public ObservableCollection<Soba> SobeListe { get; set; }
+        public INavigationService NavigationService { get; set; }
 
         public RoomsVM()
         {
-            Brisi = new RelayCommand<object>(obrisi, p => true);
-            using(var db=new HotelDbContext())
+            using (var db = new HotelDbContext())
             {
-                SobeListe = db.Sobe.OrderBy(c => c.Broj).ToList();
+                SobeListe = new ObservableCollection<Soba>(db.Sobe.OrderBy(c => c.Broj).ToList());
+            }
+            NavigationService = new NavigationService();
+            Edituj = new RelayCommand<object>(edit, p => true);
+            Brisi = new RelayCommand<object>(obrisi, p => true);
+            AddRoom = new RelayCommand<object>(dodaj, p => true);
+        }
+
+        public async void obrisi(object p)
+        {
+            if (Selektovana != null)
+            {
+                var pitanje = new MessageDialog("Are you sure you want to delete selected room?");
+                pitanje.Commands.Add(new UICommand("Yes", new UICommandInvokedHandler(this.CommandInvokedHandler)) { Id = 0 });
+                pitanje.Commands.Add(new UICommand("No", new UICommandInvokedHandler(this.CommandInvokedHandler)) { Id = 1 });
+                pitanje.DefaultCommandIndex = 0;
+                pitanje.CancelCommandIndex = 1;
+                await pitanje.ShowAsync();
+            }
+            else
+            {
+                var dialog = new MessageDialog("You haven't selected room!");
+                dialog.Title = "Error";
+                await dialog.ShowAsync();
+            }
+        }
+        private async void CommandInvokedHandler(IUICommand command)
+        {
+            if (command.Label == "Yes")
+            {
+                using (var db = new HotelDbContext())
+                {
+                    db.Sobe.Remove(Selektovana);
+                    db.SaveChanges();
+                    var dialog = new MessageDialog("You have deleted room " + Selektovana.Naziv);
+                    await dialog.ShowAsync();
+                    SobeListe.Remove(Selektovana);
+                }
             }
         }
 
-        public void obrisi(object p)
+        public void dodaj(object p)
         {
-            using (var db = new HotelDbContext())
+            NavigationService.Navigate(typeof(NewRoom));
+        }
+
+        public async void edit(object p)
+        {
+           if(Selektovana!=null) NavigationService.Navigate(typeof(EditRoom), new EditRoomVM(this));
+           else
             {
-                db.Sobe.Remove(Selektovana);
-                db.SaveChanges();
-                SobeListe = db.Sobe.OrderBy(c => c.Broj).ToList();
+                var dialog = new MessageDialog("You haven't selected room!");
+                dialog.Title = "Error";
+                await dialog.ShowAsync();
             }
         }
     }
